--- test_scripts/Polices/Policy_Table_Update/002_ATF_P_TC_PTU_AppID_NotListed_PT_DeviceConsented_SecondApp.lua
@@ -1,100 +1,0 @@
----------------------------------------------------------------------------------------------
--- Requirements summary:
--- [PolicyTableUpdate] Request PTU - an app registered is not listed in PT (device consented)
---
--- Description:
--- SDL should request PTU in case new application is registered and is not listed in PT
--- and device is consented.
--- 1. Used preconditions
--- SDL is built with "-DEXTENDED_POLICY: EXTERNAL_PROPRIETARY" flag
--- Connect mobile phone over WiFi.
--- Register new application.
--- Successful PTU. Device is consented.
--- 2. Performed steps
--- Register new application
---
--- Expected result:
--- PTU is requested. PTS is created.
--- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
--- SDL->HMI: BasicCommunication.PolicyUpdate
----------------------------------------------------------------------------------------------
-
---[[ General configuration parameters ]]
-config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
-
---[[ Required Shared libraries ]]
-local commonSteps = require('user_modules/shared_testcases/commonSteps')
-local commonFunctions = require('user_modules/shared_testcases/commonFunctions')
-local testCasesForPolicyTable = require('user_modules/shared_testcases/testCasesForPolicyTable')
-local testCasesForBuildingSDLPolicyFlag = require('user_modules/shared_testcases/testCasesForBuildingSDLPolicyFlag')
-local testCasesForPolicyTableSnapshot = require('user_modules/shared_testcases/testCasesForPolicyTableSnapshot')
-
---[[ General Precondition before ATF start ]]
--- commonFunctions:SDLForceStop()
-testCasesForBuildingSDLPolicyFlag:Update_PolicyFlag("ENABLE_EXTENDED_POLICY", "OFF")
-testCasesForBuildingSDLPolicyFlag:CheckPolicyFlagAfterBuild("ENABLE_EXTENDED_POLICY","OFF")
-commonSteps:DeleteLogsFileAndPolicyTable()
-
---ToDo: shall be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
-config.defaultProtocolVersion = 2
-
---[[ General Settings for configuration ]]
-Test = require('connecttest')
-require('cardinalities')
-require('user_modules/AppTypes')
-local mobile_session = require('mobile_session')
-
---[[ Preconditions ]]
-commonFunctions:newTestCasesGroup("Preconditions")
-testCasesForPolicyTable:flow_PTU_SUCCEESS_EXTERNAL_PROPRIETARY()
-
---[[ Test ]]
-commonFunctions:newTestCasesGroup("Test")
-function Test:TestStep_StartNewSession()
-  self.mobileSession1 = mobile_session.MobileSession( self, self.mobileConnection)
-  self.mobileSession1:StartService(7)
-end
-
-function Test:TestStep_PTU_AppID_SecondApp_NotListed_PT()
-  local hmi_app1_id = self.applications[config.application1.registerAppInterfaceParams.appName]
-
-  local correlationId = self.mobileSession1:SendRPC("RegisterAppInterface", config.application2.registerAppInterfaceParams)
-
-  EXPECT_HMINOTIFICATION("BasicCommunication.OnAppRegistered", { application = { appName = config.application2.appName } })
-  :Do(function(_,data)
-      local hmi_app2_id = data.params.application.appID
-      EXPECT_HMINOTIFICATION("SDL.OnStatusUpdate", {status = "UPDATE_NEEDED"})
-
-      testCasesForPolicyTableSnapshot:create_PTS(true, {
-          config.application1.registerAppInterfaceParams.appID,
-          config.application2.registerAppInterfaceParams.appID,
-        },
-        {config.deviceMAC},
-        {hmi_app1_id, hmi_app2_id})
-
-      local timeout_after_x_seconds = testCasesForPolicyTableSnapshot:get_data_from_PTS("module_config.timeout_after_x_seconds")
-      local seconds_between_retries = {}
-      for i = 1, #testCasesForPolicyTableSnapshot.pts_seconds_between_retries do
-        seconds_between_retries[i] = testCasesForPolicyTableSnapshot.pts_seconds_between_retries[i].value
-      end
-
-      EXPECT_HMICALL("BasicCommunication.PolicyUpdate",
-        {
-          file = "/tmp/fs/mp/images/ivsu_cache/PolicyTableUpdate",
-          timeout = timeout_after_x_seconds,
-          retry = seconds_between_retries
-        })
-      :Do(function(_,data)
-          self.hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
-        end)
-    end)
-  self.mobileSession1:ExpectResponse(correlationId, { success = true, resultCode = "SUCCESS"})
-end
-
---[[ Postconditions ]]
-commonFunctions:newTestCasesGroup("Postconditions")
-function Test:Postcondition_SDLForceStop()
-  commonFunctions:SDLForceStop(self)
-end
-
-return Test
